Windows: Fix PR13278 compile break
